+++ src/Components/Facility/ConsultationDetails.tsx
@@ -1,8 +1,10 @@
 import { navigate } from "raviger";
+import { Button, CircularProgress } from "@material-ui/core";
 import moment from "moment";
-import { useCallback, useState } from "react";
-import { useDispatch } from "react-redux";
+import React, { useCallback, useState } from "react";
+import { useDispatch, useSelector } from "react-redux";
 import { statusType, useAbortableEffect } from "../../Common/utils";
+import * as Notification from "../../Utils/Notifications";
 import {
   getConsultation,
   getDailyReport,
@@ -37,6 +39,22 @@ import TeleICUPatientInfoCard from "../TeleIcu/Patient/InfoCard";
 import TeleICUPatientVitalsCard from "../TeleIcu/Patient/VitalsCard";
 import TeleICUPatientVitalsGraphCard from "../TeleIcu/Patient/VitalsGraph";
 import DoctorVideoSlideover from "../TeleIcu/DoctorVideoSlideover";
+import { validateEmailAddress } from "../../Common/validation";
+import Dialog from "@material-ui/core/Dialog";
+import DialogActions from "@material-ui/core/DialogActions";
+import DialogContent from "@material-ui/core/DialogContent";
+import DialogContentText from "@material-ui/core/DialogContentText";
+import DialogTitle from "@material-ui/core/DialogTitle";
+import { TextInputField } from "../Common/HelperInputFields";
+import { discharge, patchPatient, dischargePatient } from "../../Redux/actions";
+
+type donatePlasmaOptionType = null | "yes" | "no" | "not-fit";
+interface preDischargeFormInterface {
+  donatePlasma: donatePlasmaOptionType;
+  disease_status?: string;
+  srf_id?: string;
+  date_of_test: any;
+}
 
 const Loading = loadable(() => import("../Common/Loading"));
 const PageTitle = loadable(() => import("../Common/PageTitle"));
@@ -49,6 +67,8 @@ export const ConsultationDetails = (props: any) => {
   const dispatch: any = useDispatch();
   const [isLoading, setIsLoading] = useState(false);
   const [showDoctors, setShowDoctors] = useState(false);
+  const state: any = useSelector((state) => state);
+  const { currentUser } = state;
 
   const [consultationData, setConsultationData] = useState<ConsultationModel>(
     {}
@@ -57,6 +77,74 @@ export const ConsultationDetails = (props: any) => {
   const [cameraAsset, setCameraAsset] = useState({});
   const [cameraMiddlewareHostname, setCameraMiddlewareHostname] = useState({});
   const [cameraConfig, setCameraConfig] = useState({});
+
+  const [open, setOpen] = React.useState(false);
+  const [openDischargeDialog, setOpenDischargeDialog] = React.useState(false);
+  const [isSendingDischargeApi, setIsSendingDischargeApi] = useState(false);
+
+  const initDischargeSummaryForm: { email: string } = {
+    email: "",
+  };
+  const [dischargeSummaryState, setDischargeSummaryForm] = useState(
+    initDischargeSummaryForm
+  );
+
+  const initErr: any = {};
+  const [errors, setErrors] = useState(initErr);
+
+  const initPreDischargeForm: preDischargeFormInterface = {
+    donatePlasma: null,
+    date_of_test: null,
+  };
+
+  const preDischargeForm = initPreDischargeForm;
+
+  const handleClickOpen = () => {
+    setOpen(true);
+  };
+
+  const handleDischageClickOpen = () => {
+    setOpenDischargeDialog(true);
+  };
+
+  const handleClose = () => {
+    setOpen(false);
+  };
+
+  const handleDischargeClose = () => {
+    setOpenDischargeDialog(false);
+  };
+
+  const handleDischargeSummarySubmit = () => {
+    if (!dischargeSummaryState.email) {
+      const errorField = Object.assign({}, errors);
+      errorField["dischargeSummaryForm"] = "email field can not be blank.";
+      setErrors(errorField);
+    } else if (!validateEmailAddress(dischargeSummaryState.email)) {
+      const errorField = Object.assign({}, errors);
+      errorField["dischargeSummaryForm"] = "Please Enter a Valid Email Address";
+      setErrors(errorField);
+    } else {
+      dispatch(
+        discharge(
+          { email: dischargeSummaryState.email },
+          { external_id: patientData.id }
+        )
+      ).then((response: any) => {
+        if ((response || {}).status === 200) {
+          Notification.Success({
+            msg: "We will be sending an email shortly. Please check your inbox.",
+          });
+        }
+      });
+      setOpen(false);
+    }
+  };
+
+  const handleDischargeSummary = (e: any) => {
+    e.preventDefault();
+    setOpen(false);
+  };
 
   const getPatientGender = (patientData: any) =>
     GENDER_TYPES.find((i) => i.id === patientData.gender)?.text;
@@ -76,6 +164,84 @@ export const ConsultationDetails = (props: any) => {
       return "None";
     }
   };
+
+  const handlePatientDischarge = async (value: boolean) => {
+    setIsSendingDischargeApi(true);
+    const dischargeData = Object.assign({}, patientData);
+    dischargeData["discharge"] = value;
+
+    // calling patchPatient and dischargePatient together caused problems check https://github.com/coronasafe/care_fe/issues/758
+
+    // using preDischargeForm form data to update patient data
+    const preDischargeFormData = formatPreDischargeFormData(preDischargeForm);
+
+    if (Object.keys(preDischargeFormData).length) {
+      // skip calling patient update api if nothing to update
+      await dispatch(
+        patchPatient(preDischargeFormData, {
+          id: patientData.id,
+        })
+      );
+    }
+    // discharge call
+    const dischargeResponse = await dispatch(
+      dischargePatient({ discharge: value }, { id: patientData.id })
+    );
+
+    setIsSendingDischargeApi(false);
+    if (dischargeResponse?.status === 200) {
+      const dischargeData = Object.assign({}, patientData);
+      dischargeData["discharge"] = value;
+      setPatientData(dischargeData);
+
+      Notification.Success({
+        msg: "Patient Discharged",
+      });
+      setOpenDischargeDialog(false);
+      window.location.reload();
+    }
+  };
+
+  const formatPreDischargeFormData = (
+    preDischargeForm: preDischargeFormInterface
+  ) => {
+    const data: any = { ...preDischargeForm };
+    const donatePlasma = preDischargeForm.donatePlasma;
+
+    if (donatePlasma) {
+      if (donatePlasma === "yes") {
+        data["will_donate_blood"] = true;
+        data["fit_for_blood_donation"] = true;
+      } else if (donatePlasma === "no") {
+        data["will_donate_blood"] = false;
+      } else if (donatePlasma === "not-fit") {
+        data["will_donate_blood"] = true;
+        data["fit_for_blood_donation"] = false;
+      }
+    }
+
+    delete data.donatePlasma;
+    return data;
+  };
+
+  const handleDischargeSummaryFormChange = (e: any) => {
+    const { value } = e.target;
+
+    const errorField = Object.assign({}, errors);
+    errorField["dischargeSummaryForm"] = null;
+    setErrors(errorField);
+
+    setDischargeSummaryForm({ email: value });
+  };
+
+  const dischargeSummaryFormSetUserEmail = () => {
+    if (!currentUser.data.email.trim())
+      return Notification.Error({
+        msg: "Email not provided! Please update profile",
+      });
+    setDischargeSummaryForm({ email: currentUser.data.email });
+  };
+
 
   const fetchData = useCallback(
     async (status: statusType) => {
@@ -168,6 +334,85 @@ export const ConsultationDetails = (props: any) => {
 
   return (
     <div>
+      <Dialog open={open} onClose={handleDischargeSummary}>
+        <DialogTitle id="form-dialog-title">
+          Download Discharge Summary
+        </DialogTitle>
+        <DialogContent>
+          <DialogContentText>
+            Please enter your email id to receive the discharge summary.
+            Disclaimer: This is an automatically Generated email using your info
+            Captured in Care System.
+            <div
+              className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative"
+              role="alert"
+            >
+              <strong className="block sm:inline font-bold">
+                Please check your email id before continuing. We cannot deliver
+                the email if the email id is invalid
+              </strong>
+            </div>
+          </DialogContentText>
+          <div className="flex justify-end">
+            <a
+              href="#"
+              className="text-xs"
+              onClick={dischargeSummaryFormSetUserEmail}
+            >
+              Fill email input with my email.
+            </a>
+          </div>
+          <TextInputField
+            type="email"
+            name="email"
+            label="email"
+            variant="outlined"
+            margin="dense"
+            autoComplete="off"
+            value={dischargeSummaryState.email}
+            InputLabelProps={{ shrink: !!dischargeSummaryState.email }}
+            onChange={handleDischargeSummaryFormChange}
+            errors={errors.dischargeSummaryForm}
+          />
+        </DialogContent>
+        <DialogActions>
+          <Button onClick={handleClose} color="primary">
+            Cancel
+          </Button>
+          <Button onClick={handleDischargeSummarySubmit} color="primary">
+            Submit
+          </Button>
+        </DialogActions>
+      </Dialog>
+      <Dialog
+        maxWidth={"md"}
+        open={openDischargeDialog}
+        onClose={handleDischargeClose}
+      >
+        <DialogContent className="px-20">
+          <div className="flex justify-center">
+            <span className="text-md text-black-800">
+              Are you sure you want to discharge {patientData.name}?
+            </span>
+          </div>
+        </DialogContent>
+        <DialogActions className="flex justify-between mt-5 px-5 border-t">
+          <Button onClick={handleDischargeClose}>Cancel</Button>
+
+          {isSendingDischargeApi ? (
+            <CircularProgress size={20} />
+          ) : (
+            <Button
+              color="primary"
+              onClick={() => handlePatientDischarge(false)}
+              autoFocus
+              // disabled={preDischargeForm.disease_status ? false : true}
+            >
+              Proceed with Discharge
+            </Button>
+          )}
+        </DialogActions>
+      </Dialog>
       <div className="px-2 pb-2">
         <nav className="flex justify-between flex-wrap">
           <PageTitle
@@ -211,55 +456,10 @@ export const ConsultationDetails = (props: any) => {
           </div>
         </nav>
         <div className="flex md:flex-row flex-col w-full mt-2">
-          <div className="border rounded-lg bg-white shadow h-full text-black p-4 w-full">
-            <div>
-              <div className="flex md:flex-row flex-col md:items-center">
-                <div className="text-2xl md:mt-2 font-semibold">
-                  <i
-                    className="text-gray-500 fas fa-hospital text-2xl"
-                    aria-hidden="true"
-                  ></i>{" "}
-                  {consultationData.facility_name}
-                </div>
-              </div>
-              <div className="flex items-center">
-                {consultationData.ip_no && (
-                  <div className="md:col-span-2 capitalize pl-2">
-                    <span className="badge badge-pill badge-primary">
-                      {`IP: ${consultationData.ip_no}`}
-                    </span>
-                  </div>
-                )}
-              </div>
-            </div>
-            <TeleICUPatientInfoCard patient={patientData} />
-            <div className="flex md:flex-row flex-col mt-4 gap-2 justify-between">
-              <div className="flex flex-col text-xs text-gray-700 font-base leading-relaxed">
-                <div>
-                  <span className="text-gray-900">Created: </span>
-                  {moment(consultationData.created_date).format("lll")} |
-                </div>
-                {consultationData.created_by && (
-                  <div>
-                    {` ${consultationData.created_by?.first_name} ${consultationData.created_by?.last_name}  `}
-                    {`@${consultationData.created_by?.username} (${consultationData.created_by?.user_type})`}
-                  </div>
-                )}
-              </div>
-              <div className="flex flex-col text-xs md:text-right text-gray-700 font-base leading-relaxed">
-                <div>
-                  <span className="text-gray-900">Last Modified: </span>
-                  {moment(consultationData.modified_date).format("lll")} |
-                </div>
-                {consultationData.last_edited_by && (
-                  <div>
-                    {` ${consultationData.last_edited_by?.first_name} ${consultationData.last_edited_by?.last_name}  `}
-                    {`@${consultationData.last_edited_by?.username} (${consultationData.last_edited_by?.user_type})`}
-                  </div>
-                )}
-              </div>
-            </div>
-            <div className="flex md:flex-row flex-col justify-between">
+          <div className="border rounded-lg bg-white shadow h-full text-black w-full">
+            <TeleICUPatientInfoCard patient={patientData} ip_no ={consultationData.ip_no}/>
+
+            <div className="flex md:flex-row flex-col justify-between border-t px-4 pt-5">
               {consultationData.admitted_to && (
                 <div className="border rounded-lg bg-gray-100 p-2 md:mt-0 mt-2">
                   <div className="border-b-2 py-1">
@@ -293,24 +493,81 @@ export const ConsultationDetails = (props: any) => {
               )}
             </div>
 
-            <div className="mt-2">
-              {consultationData.diagnosis && (
-                <div className="text-sm w-full">
-                  <span className="font-semibold leading-relaxed">
-                    Diagnosis:{" "}
-                  </span>
-                  {consultationData.diagnosis}
+            <div className="flex px-4">
+              <div className="flex-1">
+                {/*consultationData.other_symptoms && (
+                  <div className="capitalize">
+                    <span className="font-semibold leading-relaxed">
+                      Other Symptoms:{" "}
+                    </span>
+                    {consultationData.other_symptoms}
+                  </div>
+                )*/}
+
+                {consultationData.diagnosis && (
+                  <div className="text-sm w-full">
+                    <span className="font-semibold leading-relaxed">
+                      Diagnosis:{" "}
+                    </span>
+                    {consultationData.diagnosis}
+                  </div>
+                )}
+                {consultationData.verified_by && (
+                  <div className="text-sm mt-2">
+                    <span className="font-semibold leading-relaxed">
+                      Verified By:{" "}
+                    </span>
+                    {consultationData.verified_by}
+                    <i className="fas fa-check-circle fill-current text-lg text-green-500 ml-2"></i>
+                  </div>
+                )}
+              </div>
+              <div className="flex-1 text-right">
+                <button className="btn btn-primary" onClick={handleClickOpen}>
+                  <i className="fas fa-clipboard-list"></i>
+                  &nbsp;
+                  Discharge Summary
+                </button>
+
+                <button
+                  className="btn btn-primary ml-2"
+                  onClick={handleDischageClickOpen}
+                  disabled={
+                    !patientData.is_active ||
+                    patientData.last_consultation?.facility !== facilityId
+                  }
+                >
+                  <i className="fas fa-hospital-user"></i>
+                  &nbsp;
+                  Discharge from CARE
+                </button>
+              </div>
+            </div>
+            <div className="flex md:flex-row flex-col gap-2 justify-between p-4">
+              <div className="flex flex-col text-xs text-gray-700 font-base leading-relaxed">
+                <div>
+                  <span className="text-gray-900">Created: </span>
+                  {moment(consultationData.created_date).format("lll")} |
                 </div>
-              )}
-              {consultationData.verified_by && (
-                <div className="text-sm mt-2">
-                  <span className="font-semibold leading-relaxed">
-                    Verified By:{" "}
-                  </span>
-                  {consultationData.verified_by}
-                  <i className="fas fa-check-circle fill-current text-lg text-green-500 ml-2"></i>
+                {consultationData.created_by && (
+                  <div>
+                    {` ${consultationData.created_by.first_name} ${consultationData.created_by.last_name}  `}
+                    {`@${consultationData.created_by.username} (${consultationData.created_by.user_type})`}
+                  </div>
+                )}
+              </div>
+              <div className="flex flex-col text-xs md:text-right text-gray-700 font-base leading-relaxed">
+                <div>
+                  <span className="text-gray-900">Last Modified: </span>
+                  {moment(consultationData.modified_date).format("lll")} |
                 </div>
-              )}
+                {consultationData.last_edited_by && (
+                  <div>
+                    {` ${consultationData.last_edited_by.first_name} ${consultationData.last_edited_by.last_name}  `}
+                    {`@${consultationData.last_edited_by.username} (${consultationData.last_edited_by.user_type})`}
+                  </div>
+                )}
+              </div>
             </div>
           </div>
         </div>
