+++ src/Components/TeleIcu/Patient/InfoCard.tsx
@@ -1,23 +1,19 @@
 import { Link } from "raviger";
-import { GENDER } from "../../../Common/constants";
 import { getDimensionOrDash } from "../../../Common/utils";
 import { PatientModel } from "../../Patient/models";
-import { RightArrowIcon } from "../Icons/ArrowIcon";
 import { Modal } from "@material-ui/core";
 import Beds from "../../Facility/Consultations/Beds";
 import { useState } from "react";
 
-export interface ITeleICUPatientInfoCardProps {
-  patient: PatientModel;
-}
-
-export default function TeleICUPatientInfoCard({
-  patient,
-}: ITeleICUPatientInfoCardProps) {
+export default function TeleICUPatientInfoCard(props : { patient : PatientModel , ip_no? : string | undefined }) {
   const [open, setOpen] = useState(false);
+
+  const patient = props.patient;
+  const ip_no = props.ip_no;
+
   console.log(patient);
   return (
-    <section className="flex items-stretch my-2 lg:flex-row flex-col space-y-3 lg:space-y-0 lg:space-x-2">
+    <section className="flex items-start lg:flex-row flex-col space-y-3 lg:space-y-0 lg:space-x-2 justify-between">
       <Modal open={open} onClose={() => setOpen(false)}>
         <div className="bg-white md:w-4/5 p-4 mx-auto ">
           {patient?.facility &&
@@ -33,123 +29,131 @@ export default function TeleICUPatientInfoCard({
           )}
         </div>
       </Modal>
-      <div className="bg-white border-b p-5 flex items-center lg:w-7/12 w-full">
-        {patient.blood_group && (
-          <div className="flex flex-col items-center ">
-            <div className="text-2xl border border-gray-300 rounded-full font-semibold items-center m-2 h-20 w-20 flex justify-center">
-              {patient.blood_group === "UNKNOWN" ? "?" : patient.blood_group}
-            </div>
-            <span className="font-light text-primary-600 text-xs mr-1">
-              Blood Group
-            </span>
-          </div>
-        )}
-        <div className="pl-4">
-          <p className="sm:text-xl md:text-4xl font-semibold ml-1">
+      <div className="bg-white p-6 flex lg:w-7/12 w-full">
+
+        {/* Can support for patient picture in the future */}
+        <div className="text-center">
+          <div className="w-24 h-24 rounded-2xl bg-primary-100 text-5xl flex justify-center items-center">
+            {!patient.last_consultation?.current_bed ? (
+              <i className="fas fa-user-injured text-primary-600"></i>
+            ) : (
+              <span className="text-2xl text-primary-600 font-semibold">
+                {patient.last_consultation?.current_bed?.bed_object?.name}
+              </span>
+            )}
+            
+          </div> 
+          <button
+            className="text-sm text-primary-600 hover:bg-gray-300 p-2 rounded m-1"
+            onClick={() => setOpen(true)}
+          >
+            {!patient.last_consultation?.current_bed ? "Assign Bed" : "Switch Bed"}
+          </button>
+        </div>
+        <div className="pl-6">
+          <p className="sm:text-xl md:text-4xl font-bold mb-1">
             {patient.name}
           </p>
+          <div className="flex align-center mb-2">
+            <Link
+              href={`/facility/${patient.facility_object?.id}`}
+              className="font-semibold text-black hover:text-primary-600"
+            >
+              <i className="text-primary-400 fas fa-hospital mr-1" aria-hidden="true"></i>
+              {patient.facility_object?.name}
+            </Link>
+            {ip_no && (
+              <span className="md:col-span-2 capitalize pl-2">
+                <span className="badge badge-pill badge-primary">
+                  {`IP: ${ip_no}`}
+                </span>
+              </span>
+            )}
+          </div>
           {!patient.is_active && (
-            <div className="bg-red-400 text-white inline-block rounded-lg px-2 py-1">
+            <div className="bg-red-200 text-red-600 inline-block mb-3 rounded-lg px-2 py-1">
               Discharged from CARE
             </div>
           )}
-          <p className="text-sm sm:text-base my-1 ml-1 text-primary-800">
+          <p className="text-sm sm:text-sm text-gray-900">
             <span>{patient.age} years</span>
             <span className="mx-2">•</span>
-            <span>{patient.gender && GENDER[patient.gender]}</span>
+            <span>{patient.gender}</span>
           </p>
-          <div className="text-sm flex flex-wrap items-center">
-            {!patient.last_consultation?.current_bed ? (
-              <button
-                className="text-sm text-primary-600 hover:bg-gray-300 p-2 rounded"
-                onClick={() => setOpen(true)}
-              >
-                Assign Bed
-              </button>
-            ) : (
-              <div className="m-1">
-                <span className="sm:text-base font-semibold text-sm mr-2">
-                  {patient.last_consultation?.current_bed?.bed_object?.name}
-                </span>
-                <button
-                  className="text-sm text-primary-600 hover:bg-gray-300 p-2 rounded"
-                  onClick={() => setOpen(true)}
-                >
-                  Switch Bed
-                </button>
-              </div>
-            )}
-            {(patient.last_consultation?.weight || false) && (
-              <div className="m-1">
-                <span className="font-light text-primary-600 text-sm mr-1">
-                  Weight
-                </span>
-                <span className="sm:text-base font-semibold text-sm mr-2">
-                  {getDimensionOrDash(patient.last_consultation.weight, " kg")}
-                </span>
-              </div>
-            )}
-            {(patient.last_consultation?.height || false) && (
-              <div className="m-1">
-                <span className="font-light text-primary-600 text-sm mr-1">
-                  Height
-                </span>
-                <span className="sm:text-base font-semibold text-sm mr-2">
-                  {getDimensionOrDash(patient.last_consultation?.height, "cm")}
-                </span>
-              </div>
-            )}
+          <div className="text-sm flex flex-wrap gap-2 mt-4">
+
+            {
+              [
+                [
+                  "Blood Group",
+                  patient.blood_group,
+                  patient.blood_group,
+                ],
+                [
+                  "Weight",
+                  getDimensionOrDash(patient.last_consultation?.weight, " kg"),
+                  true
+                ],
+                [
+                  "Height",
+                  getDimensionOrDash(patient.last_consultation?.height, "cm"),
+                  true
+                ]
+              ].map((stat, i)=>{
+                return stat[2] ? (
+                  <div key={`patient_stat_`+i} className="bg-gray-200 border-gray-500 border py-1 px-2 rounded-lg text-xs">
+                    <b>
+                      {stat[0]}
+                    </b> : {stat[1]}
+                  </div>
+                ) : ""
+              })
+            }
           </div>
         </div>
       </div>
       <div
-        className={`grid grid-cols-${
-          patient.is_active ? "2" : "1"
-        } gap-x-4 flex-1 gap-y-2`}
+        className={`flex gap-2 flex-col bg-gray-100 p-6 `}
       >
-        {patient.is_active &&
-          patient.last_consultation?.id &&
+        {
           [
-            ["update", "Edit Consultation Details", "pencil-alt"],
-            ["daily-rounds", "Log Update", "plus"],
-          ].map((action, i) => (
+            [
+              `/facility/${patient.facility}/patient/${patient.id}/consultation/${patient.last_consultation?.id}/update`, 
+              "Edit Consultation Details",
+              "pencil-alt", 
+              patient.is_active && patient.last_consultation?.id
+            ],
+            [
+              `/facility/${patient.facility}/patient/${patient.id}/consultation/${patient.last_consultation?.id}/daily-rounds`,
+              "Log Update", 
+              "plus", 
+              patient.is_active && patient.last_consultation?.id
+            ],
+            [
+              `/patient/${patient.id}/investigation_reports`, 
+              "Investigation Summary", 
+              "address-card", 
+              true
+            ],
+            [
+              `/facility/${patient.facility}/patient/${patient.id}/consultation/${patient.last_consultation?.id}/treatment-summary`,
+              "Treatment Summary", 
+              "prescription-bottle-medical", 
+              patient.last_consultation?.id
+            ]
+          ].map((action, i) => action[3] && (
             <Link
               key={i}
-              href={`/facility/${patient.facility}/patient/${patient.id}/consultation/${patient.last_consultation?.id}/${action[0]}`}
-              className="rounded-md bg-white shadow-sm text-primary-900 flex justify-between items-center p-2 px-4 sm:px-2 hover:bg-gray-200 cursor-pointer active:translate-y-1 transform"
+              href={`${action[0]}`}
+              className="btn btn-primary justify-start hover:text-white"
             >
-              <p className="text-sm sm:text-base px-2 font-semibold">
+              <i className={`fas fa-${action[2]} w-4 mr-3`}></i>
+              <p className="font-semibold">
                 {action[1]}
               </p>
-              <span>
-                <i className={`fas fa-${action[2]} mr-2`}></i>
-              </span>
             </Link>
-          ))}
-        <Link
-          href={`/patient/${patient.id}/investigation_reports`}
-          className="rounded-md bg-white shadow-sm text-primary-900 flex justify-between items-center p-2 px-4 sm:px-2 hover:bg-gray-200 cursor-pointer active:translate-y-1 transform"
-        >
-          <p className="text-sm sm:text-base px-2 font-semibold">
-            Investigation Summary
-          </p>
-          <span>
-            <RightArrowIcon />
-          </span>
-        </Link>
-        {patient.last_consultation?.id && (
-          <Link
-            href={`/facility/${patient.facility}/patient/${patient.id}/consultation/${patient.last_consultation?.id}/treatment-summary`}
-            className="rounded-md bg-white shadow-sm text-primary-900 flex justify-between items-center p-2 px-4 sm:px-2 hover:bg-gray-200 cursor-pointer active:translate-y-1 transform"
-          >
-            <p className="text-sm sm:text-base px-2 font-semibold">
-              Treatment Summary
-            </p>
-            <span>
-              <RightArrowIcon />
-            </span>
-          </Link>
-        )}
+          ))
+        }
       </div>
     </section>
   );
